+++ snake.py
@@ -7,6 +7,11 @@ import signal
 import multiprocessing
 from playsound import playsound
 import requests
+import logging
+# import logging.handlers
+import sys
+import json
+import base64
 
 
 ######## Classes and definitions ########
@@ -36,6 +41,18 @@ def Pkill(process_name):
         process_name.terminate()
 
 
+def base64_encode(text):
+    encodedtext_bytes = base64.b64encode(text.encode("utf-8"))
+    encodedtext = str(encodedtext_bytes, "utf-8")
+    return encodedtext
+
+
+def base64_decode(text):
+    decodedtext_bytes = base64.b64decode(text)
+    decodedtext = str(decodedtext_bytes, "utf-8")
+    return decodedtext
+
+
 class Button:
     def __init__(self, x, y, width, height, color1, color2, color_text, text, font_size, command):
         self.x = x
@@ -45,7 +62,7 @@ class Button:
         self.button_color1 = color1
         self.button_color2 = color2
         self.color_text = color_text
-        self.font = pygame.font.SysFont("Verdana", font_size, bold=True)
+        self.font = pygame.font.SysFont(settings.font_name, font_size, bold=True)
         self.text_width, self.text_height = self.font.size(text)
         self.text = self.font.render(text, True, color_text)
         self.command = command
@@ -64,7 +81,7 @@ class Button:
 
 class Text:
     def __init__(self, text, color, font_size):
-        self.font = pygame.font.SysFont("Verdana", font_size, bold=True)
+        self.font = pygame.font.SysFont(settings.font_name, font_size, bold=True)
         self.text_width, self.text_height = self.font.size(text)
         self.text = self.font.render(text, True, color)
 
@@ -72,45 +89,135 @@ class Text:
         window.blit(self.text, (x, y))
 
 
+class LongText:
+    def __init__(self, text, color, font_size, line_lenght, line_spacing):
+        self.line_spacing = line_spacing
+        self.font = pygame.font.SysFont(settings.font_name, font_size, bold=True)
+
+        # splitting text into lines
+        words = text.split(" ")
+        words_len = []
+        for word in words:
+            words_len.append(len(word))
+
+        i = 0
+        textList = []
+        curr_line = ""
+        while i < len(words):
+            if line_lenght - len(curr_line) - 1 >= words_len[i]:
+                if len(curr_line):
+                    curr_line += " "
+                curr_line += words[i]
+                i += 1
+
+            elif words_len[i] >= line_lenght:
+                textList.append(curr_line)
+                textList.append(words[i])
+                curr_line = ""
+                i += 1
+
+            else:
+                textList.append(curr_line)
+                curr_line = ""
+
+        if len(curr_line):
+            textList.append(curr_line)
+
+        #
+
+        self.rendredTextList = []
+        for line in textList:
+            self.rendredTextList.append(Text(line, color, font_size))
+
+        self.text_width = self.rendredTextList[0].text_width
+        for line in self.rendredTextList:
+            if line.text_width > self.text_width:
+                self.text_width = line.text_width
+        self.line_height = self.rendredTextList[0].text_height
+        self.text_height = len(self.rendredTextList) * self.line_height + (len(self.rendredTextList) - 1) * line_spacing
+
+    def draw(self, x, y):
+        i = 0
+        for line in self.rendredTextList:
+            line.draw(x, y + i)
+            i += self.line_height + self.line_spacing
+
+
 class File:  # Data
     def __init__(self):
-        self.path = settings.path_highscore
+        self.path_data = settings.path_data
 
     def read(self):
-        file = open(self.path, "r")
-        line = file.readline()
-        if line == "":
-            self.highscore = 0
-        else:
-            self.highscore = int(line)
+        try:
+            with open(self.path_data, "r") as file:
+                data = base64_decode(file.readline())
 
-        file.close()
+            self.datadict = json.loads(data)
+            if not self.datadict.get("highscore"):
+                self.highscore = 0
+            else:
+                self.highscore = self.datadict.get("highscore")
+
+        except Exception as e:
+            logger.error("Error while reading game data - type: " + str(type(e)))
+            logger.error("Error while reading game data: " + str(e))
+            error_screen("Error while reading game data")
+
+        else:
+            logger.info("Game data successfully read")
 
     def write(self):
-        self.highscore = Snake.score
-        file = open(self.path, "w")
-        file.write(str(Snake.score))
+        try:
+            self.datadict["highscore"] = self.highscore
+            with open(self.path_data, "w") as file:
+                file.write(base64_encode(json.dumps(self.datadict)))
+        except Exception as e:
+            logger.error("Error while writing game data - type: " + str(type(e)))
+            logger.error("Error while writing game data: " + str(e))
+            error_screen("Error while writing game data")
 
 
 def checkFiles():
-    if not os.path.exists(settings.path_gameDirectory):
-        os.mkdir(settings.path_gameDirectory)
-
-    if not os.path.exists(settings.path_highscore):
-        file = open(settings.path_highscore, "a")
-        file.close()
+    if not os.path.exists(settings.path_data) or os.path.getsize(settings.path_data) == 0:
+        logger.warning("Data file did not exist, trying to create")
+        with open(settings.path_data, "w") as file:
+            empty_dict = {
+                "highscore": None,
+                "just found an Easter Egg?": True
+            }
+            file.write(base64_encode(json.dumps(empty_dict)))
+        logger.warning("Data file successfully created")
 
     if not os.path.exists(settings.path_musicDirectory):
+        logger.warning("Music directory did not exist, trying to create")
         os.mkdir(settings.path_musicDirectory)
+        logger.warning("Music directory successfully created")
 
     if not os.path.exists(settings.path_music_Game):
-        download = requests.get(settings.url_music_Game, allow_redirects=True)
-        open(settings.path_music_Game, "wb").write(download.content)
+        logger.warning("Game music did not exist, trying to download")
+        try:
+            download = requests.get(settings.url_music_Game, allow_redirects=True)
+            open(settings.path_music_Game, "wb").write(download.content)
+        except Exception as e:
+            logger.error("Downloading error type: " + str(type(e)))
+            logger.error("Downloading error: " + str(e))
+            exit(1)
+        else:
+            logger.warning("Game music successfully downloaded")
 
     if not os.path.exists(settings.path_music_GameOver):
-        download = requests.get(settings.url_music_GameOver, allow_redirects=True)
-        open(settings.path_music_GameOver, "wb").write(download.content)
-    print("Checking files done")
+        logger.warning("GameOver music did not exist, trying to download")
+        try:
+            download = requests.get(settings.url_music_GameOver, allow_redirects=True)
+            open(settings.path_music_GameOver, "wb").write(download.content)
+        except Exception as e:
+            logger.error("Downloading error type: " + str(type(e)))
+            logger.error("Downloading error: " + str(e))
+            exit(1)
+        else:
+            logger.warning("GameOver music successfully downloaded")
+
+    logger.info("Checking files done")
 
 
 def loading_screen(function, text):
@@ -136,6 +243,34 @@ def loading_screen(function, text):
             Loading3.draw((settings.window_width - Loading3.text_width) / 2, (settings.window_height - Loading3.text_height) / 2)
         pygame.display.update()
         time += 1
+
+    if process.exitcode:
+        error_screen("Program encountered a problem while creating local files. Check Your Internet connection and try again later.")
+
+
+def error_screen(text):
+    logger.critical("Error screen: " + text)
+    global error
+    global mouse
+    error = True
+    ErrorText = LongText(text, settings.color_font, settings.font_size_error, settings.line_lenght_error, settings.line_spacing)
+    ButtonExit2 = Button(int((settings.window_width - settings.button_width) / 2), 500, settings.button_width, settings.button_height, settings.color_button, settings.color_button_focused, settings.button_text_color, "Exit", settings.button_text_size, "exit(1)")
+
+    while error:
+        clock.tick(settings.fps)
+
+        keys = pygame.key.get_pressed()
+        mouse = pygame.mouse.get_pos()
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT or keys[pygame.K_ESCAPE]:
+                error = False
+            if event.type == pygame.MOUSEBUTTONDOWN:
+                ButtonExit2.click()
+
+        window.fill(settings.color_error_backgorund)
+        ErrorText.draw((settings.window_width - ErrorText.text_width) / 2, (settings.window_height - ErrorText.text_height) / 2 - 60)
+        ButtonExit2.draw()
+        pygame.display.update()
 
 
 class Player:  # Snake
@@ -242,13 +377,16 @@ class Bar:  # TopBar
 def gameOverText():
     GameOver.draw((settings.window_width - GameOver.text_width) / 2, (settings.window_height - GameOver.text_height) / 2)
     if Snake.score > Data.highscore:
+        logger.info("Highscore beaten, old: {}, new: {}".format(Data.highscore, Snake.score))
+        Data.highscore = Snake.score
         Data.write()
-        NewHigscoreText = Text("new highscore: " + str(Snake.score), settings.color_newhighscore, settings.font_size_newhighscore)
-        NewHigscoreText.draw((settings.window_width - NewHigscoreText.text_width) / 2, (settings.window_height - GameOver.text_height) / 2 - GameOver.text_height + NewHigscoreText.text_height - 10)
+        NewHighscoreText = Text("new highscore: " + str(Snake.score), settings.color_newhighscore, settings.font_size_newhighscore)
+        NewHighscoreText.draw((settings.window_width - NewHighscoreText.text_width) / 2, (settings.window_height - GameOver.text_height) / 2 - GameOver.text_height + NewHighscoreText.text_height - 10)
     pygame.display.update()
 
 
 def music_Game():
+    logger.debug("Starting game music")
     while True:
         playsound(settings.path_music_Game)
 
@@ -332,10 +470,10 @@ def game_main():
         if Snake.fpsCounter % settings.move_delay == 0:
             Snake.move()
 
+    logger.info("Game over, score: " + str(Snake.score))
     Pkill(musicGame)
     gameOverText()
     playsound(settings.path_music_GameOver)
-    # pygame.time.wait(3650)  # unnecessary, because playing sound stops code execution
 
 
 def game_redraw():
@@ -365,23 +503,29 @@ class settings:
     game_height = window_height - label_height - grid  # odd multiples of grid
 
     color_window_background = (1, 170, 64)
+    color_error_backgorund = (208, 26, 26)
     color_game_background = (0, 0, 0)
     color_snake_head = (255, 255, 255)
     color_snake_tail = (0, 255, 0)
     color_apple = (255, 0, 0)
     color_button = (254, 151, 12)
-    color_button_pointed = (183, 111, 15)
+    color_button_focused = (183, 111, 15)
     color_font = (255, 255, 255)
     color_gameover = (255, 0, 0)
     color_newhighscore = (0, 0, 255)
     color_snakeLogo = (255, 255, 255)
     color_author = (200, 200, 200)
 
+    font_name = "Verdana"
     font_size_loading = 35
+    font_size_error = 30
     font_size_gameover = 70
     font_size_newhighscore = 33
     font_size_snakeLogo = 60
     font_size_author = 21
+
+    line_spacing = 6
+    line_lenght_error = 40
 
     button_width = grid * 10
     button_height = grid * 4
@@ -391,11 +535,12 @@ class settings:
     ButtonPlay_y = 245
     ButtonExit_y = 415
 
-    path_gameDirectory = os.path.expanduser("~/.snake")
-    path_highscore = path_gameDirectory + "/.snake_data"  # os.path.expanduser("~/.snake/.snake_data")
-    path_musicDirectory = path_gameDirectory + "/music"  # os.path.expanduser("~/.snake/music")
-    path_music_Game = path_musicDirectory + "/Tristan Lohengrin - Happy 8bit Loop 01.mp3"  # os.path.expanduser("~/.snake/music/Tristan Lohengrin - Happy 8bit Loop 01.mp3")
-    path_music_GameOver = path_musicDirectory + "/Sad Trombone Wah Wah Wah Fail Sound Effect.mp3"  # os.path.expanduser("~/.snake/music/Sad Trombone Wah Wah Wah Fail Sound Effect.mp3")
+    path_gameDirectory = os.path.join(os.path.expanduser("~"), ".snake")  # ~/.snake
+    path_data = os.path.join(path_gameDirectory, "data")  # ~/.snake/data
+    path_musicDirectory = os.path.join(path_gameDirectory, "music")  # ~/.snake/music
+    path_music_Game = os.path.join(path_musicDirectory, "Tristan Lohengrin - Happy 8bit Loop 01.mp3")
+    path_music_GameOver = os.path.join(path_musicDirectory, "Sad Trombone Wah Wah Wah Fail Sound Effect.mp3")
+    path_logDirectory = os.path.join(path_gameDirectory, "logs")  # ~/.snake/logs
 
     url_music_Game = "https://drive.google.com/u/0/uc?id=12iQh-5UzBuTLsWbTAeHemYWlmwHm0USr&export=download"
     url_music_GameOver = "https://drive.google.com/u/0/uc?id=12SbYvszlHOUjhjiWk4bEo1pjqkjTaI1y&export=download"
@@ -407,11 +552,43 @@ class settings:
 
 ############# Main code #############
 
+#### Logging configuration ####
+if not os.path.exists(settings.path_gameDirectory):  # moved from checkFiles() to make sure there is a game directory and the log file can be put there
+    os.mkdir(settings.path_gameDirectory)
+if not os.path.exists(settings.path_logDirectory):
+    os.mkdir(settings.path_logDirectory)
+
+if os.path.exists(os.path.join(settings.path_logDirectory, "4.log")):
+    os.remove(os.path.join(settings.path_logDirectory, "4.log"))
+if os.path.exists(os.path.join(settings.path_logDirectory, "3.log")):
+    os.rename(os.path.join(settings.path_logDirectory, "3.log"), os.path.join(settings.path_logDirectory, "4.log"))
+if os.path.exists(os.path.join(settings.path_logDirectory, "2.log")):
+    os.rename(os.path.join(settings.path_logDirectory, "2.log"), os.path.join(settings.path_logDirectory, "3.log"))
+if os.path.exists(os.path.join(settings.path_logDirectory, "1.log")):
+    os.rename(os.path.join(settings.path_logDirectory, "1.log"), os.path.join(settings.path_logDirectory, "2.log"))
+
+for handler in logging.root.handlers[:]:  # this is needed in PyCharm and can be left for safety
+    logging.root.removeHandler(handler)
+
+logger = logging.getLogger()
+logger.setLevel(logging.DEBUG)
+formatter = logging.Formatter('%(asctime)s [%(levelname)s] (Line %(lineno)d in %(funcName)s) - %(message)s')
+file_handler = logging.FileHandler(filename=os.path.join(settings.path_logDirectory, "1.log"))
+file_handler.setFormatter(formatter)
+logger.addHandler(file_handler)
+# file_rot_handler = logging.handlers.RotatingFileHandler(filename=settings.path_log, maxBytes=60000, backupCount=1)
+# file_rot_handler.setFormatter(formatter)
+# logger.addHandler(file_rot_handler)
+stdout_handler = logging.StreamHandler(sys.stdout)
+logger.addHandler(stdout_handler)
+
+#### Main game code ####
+logger.info("Starting")
 pygame.display.init()
 pygame.font.init()
 clock = pygame.time.Clock()
 window = pygame.display.set_mode((settings.window_width, settings.window_height))
-pygame.display.set_caption("Snake by Michał v0.5")
+pygame.display.set_caption("Snake v0.6")
 
 loading_screen(checkFiles, "Loading")
 
@@ -421,8 +598,8 @@ GameOver = Text("GAME  OVER", settings.color_gameover, settings.font_size_gameov
 SnakeLogo = Text("Snake Game", settings.color_snakeLogo, settings.font_size_snakeLogo)
 Author = Text("Michał Machnikowski 2021", settings.color_author, settings.font_size_author)
 
-ButtonPlay = Button(int((settings.window_width - settings.button_width) / 2), settings.ButtonPlay_y, settings.button_width, settings.button_height, settings.color_button, settings.color_button_pointed, settings.button_text_color, "Play", settings.button_text_size, "game_main()")
-ButtonExit = Button(int((settings.window_width - settings.button_width) / 2), settings.ButtonExit_y, settings.button_width, settings.button_height, settings.color_button, settings.color_button_pointed, settings.button_text_color, "Exit", settings.button_text_size, "menu = False")
+ButtonPlay = Button(int((settings.window_width - settings.button_width) / 2), settings.ButtonPlay_y, settings.button_width, settings.button_height, settings.color_button, settings.color_button_focused, settings.button_text_color, "Play", settings.button_text_size, "game_main()")
+ButtonExit = Button(int((settings.window_width - settings.button_width) / 2), settings.ButtonExit_y, settings.button_width, settings.button_height, settings.color_button, settings.color_button_focused, settings.button_text_color, "Exit", settings.button_text_size, "menu = False")
 
 Snake = Player(settings.grid, settings.grid_border, settings.color_snake_head, settings.color_snake_tail)
 Apple = Target(settings.grid, settings.grid_border, settings.color_apple)
@@ -431,3 +608,5 @@ TopBar = Bar(settings.label_x, settings.label_y, settings.label_width, settings.
 menu_main()
 
 pygame.quit()
+
+logger.info("Quitting")